+++ resources/js/app.js
@@ -9,6 +9,11 @@ require('./bootstrap');
 window.Vue = require('vue').default;
 
 require('./popover');
+require('./modal');
+require('./resetfilter');
+require('./select2');
+require('./messger');
+require('./menu');
 
 /**
  * The following block of code may be used to automatically register your
