+++ resources/js/messger.js
@@ -1,0 +1,8 @@
+$(document).ready(function () {
+    $(".close-button").click(function () {
+        $(".text-mess").hide();
+    });
+    $(".img-mess").click(function () {
+        $(".text-mess").show();
+    });
+});
