+++ resources/js/select2.js
@@ -1,0 +1,6 @@
+$(function () {
+    $('.js-states').select2({
+        placeholder: 'Select an option',
+        theme: 'bootstrap4',
+    });
+});
